fix test

+++ internal/repository/test/bot_repo_test.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
-	"github.com/vladislavprovich/TG-bot/internal/repository"
+	repository "github.com/vladislavprovich/TG-bot/internal/repository"
 	mocks "github.com/vladislavprovich/TG-bot/mocks"
 	"go.uber.org/mock/gomock"
 	"testing"
@@ -15,13 +15,13 @@ func TestSaveURL(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	mockDB := mocks.NewMockDBExecutor(ctrl)
+	mockDB := mocks.NewMockExtContext(ctrl)
 
 	mockDB.EXPECT().ExecContext(gomock.Any(), "INSERT INTO urls (user_id, original_url, short_url) VALUES (?, ?, ?)",
 		"user123", "http://example.com", "http://short.url").Return(nil, nil)
 	logger := logrus.New()
 	repo := repository.NewBotRepository(mockDB, *logger)
-	err := repo.SaveURL(context.TODO(), repository.SaveUrlRequest{
+	err := repo.SaveURL(context.TODO(), &repository.SaveUrlRequest{
 		UserID: "user123",
 		URL: &repository.URLCombined{
 			OriginalURL: "http://example.com",
@@ -35,7 +35,7 @@ func TestGetListURL(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	mockDB := mocks.NewMockDBExecutor(ctrl)
+	mockDB := mocks.NewMockExtContext(ctrl)
 
 	mockDB.EXPECT().QueryContext(gomock.Any(), "SELECT original_url, short_url FROM urls WHERE user_id = ?", "user123").
 		Return([]*repository.URLCombined{
@@ -45,7 +45,9 @@ func TestGetListURL(t *testing.T) {
 	logger := logrus.New()
 	repo := repository.NewBotRepository(mockDB, *logger)
 
-	_, err := repo.GetListURL(context.TODO(), "user123")
+	_, err := repo.GetListURL(context.TODO(),
+		&repository.GetListURLRequest{UserID: "user123"},
+	)
 
 	require.NoError(t, err)
 }
@@ -54,7 +56,7 @@ func TestDeleteAllURL(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	mockDB := mocks.NewMockDBExecutor(ctrl)
+	mockDB := mocks.NewMockExtContext(ctrl)
 
 	mockDB.EXPECT().ExecContext(gomock.Any(), "DELETE FROM urls WHERE user_id = ?", "user123").Return(nil, nil)
 
@@ -62,7 +64,7 @@ func TestDeleteAllURL(t *testing.T) {
 
 	repo := repository.NewBotRepository(mockDB, *logger)
 
-	err := repo.DeleteAllURL(context.TODO(), "user123")
+	err := repo.DeleteAllURL(context.TODO(), &repository.DeleteAllURLRequest{UserID: "user123"})
 
 	require.NoError(t, err)
 }
@@ -70,7 +72,7 @@ func TestDeleteAllURL(t *testing.T) {
 func TestDeleteURL(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
-	mockDB := mocks.NewMockDBExecutor(ctrl)
+	mockDB := mocks.NewMockExtContext(ctrl)
 
 	mockDB.EXPECT().ExecContext(gomock.Any(), "DELETE FROM urls WHERE user_id = ? AND original_url = ?", "user123", "http://example.com").Return(nil, nil)
 
@@ -78,7 +80,7 @@ func TestDeleteURL(t *testing.T) {
 
 	repo := repository.NewBotRepository(mockDB, *logger)
 
-	err := repo.DeleteURL(context.TODO(), "user123", "http://example.com")
+	err := repo.DeleteURL(context.TODO(), &repository.DeleteURLRequest{UserID: "user123", OriginalURL: "http://example.com"})
 
 	require.NoError(t, err)
 }
@@ -87,7 +89,7 @@ func TestDeleteAllUrl_Error(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
-	mockDB := mocks.NewMockDBExecutor(ctrl)
+	mockDB := mocks.NewMockExtContext(ctrl)
 
 	mockDB.EXPECT().ExecContext(gomock.Any(), "DELETE FROM urls WHERE user_id = ?", "user123").Return(nil, errors.New("database error"))
 
@@ -95,7 +97,7 @@ func TestDeleteAllUrl_Error(t *testing.T) {
 
 	repo := repository.NewBotRepository(mockDB, *logger)
 
-	err := repo.DeleteAllURL(context.TODO(), "user123")
+	err := repo.DeleteAllURL(context.TODO(), &repository.DeleteAllURLRequest{UserID: "user123"})
 
 	require.Error(t, err)
 }
@@ -103,7 +105,7 @@ func TestDeleteAllUrl_Error(t *testing.T) {
 func TestDeleteURL_Error(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
-	mockDB := mocks.NewMockDBExecutor(ctrl)
+	mockDB := mocks.NewMockExtContext(ctrl)
 
 	mockDB.EXPECT().ExecContext(gomock.Any(), "DELETE FROM urls WHERE user_id = ? AND original_url = ?", "user123", "http://example.com").Return(nil, errors.New("database error"))
 
@@ -111,7 +113,7 @@ func TestDeleteURL_Error(t *testing.T) {
 
 	repo := repository.NewBotRepository(mockDB, *logger)
 
-	err := repo.DeleteURL(context.TODO(), "user123", "http://example.com")
+	err := repo.DeleteURL(context.TODO(), &repository.DeleteURLRequest{UserID: "user123", OriginalURL: "http://example.com"})
 
 	require.Error(t, err)
 }
